+++ Example/ThirdPartyMailerExample/ViewController.swift
@@ -3,7 +3,7 @@
 //  ThirdPartyMailerExample
 //
 //  Created by Vincent Tourraine on 28/03/16.
-//  Copyright © 2016-2020 Vincent Tourraine. All rights reserved.
+//  Copyright © 2016-2021 Vincent Tourraine. All rights reserved.
 //
 
 import UIKit
@@ -12,7 +12,6 @@ import ThirdPartyMailer
 class ViewController: UITableViewController {
 
     let clients = ThirdPartyMailClient.clients()
-
 
     // MARK: - Table view data source
 
@@ -23,11 +22,9 @@ class ViewController: UITableViewController {
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: "Cell", for: indexPath)
         let client = clients[indexPath.row]
-        let application = UIApplication.shared
-
         cell.textLabel?.text = client.name
 
-        if ThirdPartyMailer.application(application, isMailClientAvailable: client) {
+        if ThirdPartyMailer.application(isMailClientAvailable: client) {
             cell.detailTextLabel?.text = NSLocalizedString("Available", comment: "")
             cell.detailTextLabel?.textColor = view.tintColor
         }
@@ -44,9 +41,8 @@ class ViewController: UITableViewController {
 
     override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         let client = clients[indexPath.row]
-        let application = UIApplication.shared
 
-        ThirdPartyMailer.application(application, openMailClient: client, recipient: nil, subject: NSLocalizedString("Test ThirdPartyMailer", comment: ""), body: nil)
+        ThirdPartyMailer.application(openMailClient: client, recipient: nil, subject: NSLocalizedString("Test ThirdPartyMailer", comment: ""), body: nil)
 
         tableView.deselectRow(at: indexPath, animated: true)
     }